Make invoking of GHDL executable GHDL-version-specific
Some of the options, like -o, work in one GHDL version, but not
another. Create a GhdlRunner base class, implementing the basic
functionality, but customize it with children classes for each
specific version. The GhdlRunnerFactory is responsible for creating
an instance of GhdlRunner, based on the configured GHDL version.

Note that, classes handling GHDL version 0 and 1 should be complete,
as it was available previously; however, classes for version 2 and 3
are created, but yet not implemented correctly (probably).

Create a config file in user's config directory (e.g. .config on
Linux), under parvaj, named config.json. Create a new ghdl.version
key for managing GHDL version, which should be both user-configurable
and auto-detectable when not provided.