Fix AppImage build script to remove bad bits from resulting AppImage
In some Docker containers, running the AppImage does not give us the
error of FUSE lib being unavailable, but a strange "No such file or
directory". Although running inside Docker is different from a real
GNU/Linux distro and should not be a concern, but to be on the safe
side, we add this little hack.

See the following link for more information:
https://github.com/AppImage/AppImageKit/issues/828#issuecomment-731895751