Fix namespace resolution for Parvaj cache
Each project should have used a namespace equals to the path of
the root of that project. However, because of a wrong variable
name, this was mistaken. Fix this.

In order to do this, we need to replace all characters except valid
ones in the path to be able to put it as the namespace. Replace
them with a really rare and special string '_+.+_'.